Update fiio-k3-fix - pidof bug
Fixes an issue where, if multiple instances of pipewire are running, multiple PIDs get returned. This causes the username of the two PIDs to then be retrieved consecutively, which breaks the systemctl command.

The -s argument makes pidof only return the first PID it comes across, fixing the issue. This is a bad solution, but it is 1AM.